+++ src/Tgraph/Grid.hs
@@ -7,23 +7,24 @@ Maintainer  : chrisreade@mac.com
 Stability   : experimental
 
 This module defines a Grid type recording 2D points with fast lookup by grid coordinates.
-It is used to quickly identify touching vertices when forcing.
+It is used to quickly identify touching vertices (e.g. when forcing).
 -}
 {-# LANGUAGE NoMonomorphismRestriction #-}
 {-# LANGUAGE FlexibleContexts          #-}
 {-# LANGUAGE TypeFamilies              #-}
 {-# LANGUAGE TupleSections             #-}
 
+
 module Tgraph.Grid
 (
     insertGridCheck
   , createGridCheck
   , createGrid
-  , createPointGrid
   , emptyGrid
   , Grid
   , touching
-  
+  , ValuedPoint
+
 ) where
 
 import Diagrams.Prelude
@@ -33,43 +34,56 @@ import Data.List ( find )
 import Data.IntMap.Strict (IntMap)
 import qualified Data.IntMap.Strict as IMap (alter,insert,empty,lookup)
 
--- | A (Grid a) maps from Int x and y coordinates to a list of valued points in the unit square
--- with bottom left corner given by the coordinates, each valued point is a pair with a value and point.
--- For a Grid with just points (and no associated values), use Grid ().
-newtype Grid a = Grid{ gridmap::IntMap (IntMap [(a,P2 Double)])} deriving (Show)
+-- | A (Grid a) maps from Int x and y coordinates to a list of a in the unit square
+-- with bottom left corner given by the coordinates of a.
+-- The type a should be a ValuedPoint for Grid access functions to work.
+newtype Grid a = Grid{ gridmap::IntMap (IntMap [a])} deriving (Show)
 
 -- | An empty grid.
 emptyGrid :: Grid a
 emptyGrid = Grid IMap.empty
 
--- | get the Int x and Int y coordinates of a 2D point (using floor).
-gridCoords :: P2 Double -> (Int,Int)
-gridCoords p = (floor x, floor y) 
-               where x = p ^. _x 
-                     y = p ^. _y 
+-- |A class for items which have a 2D point (so can be added to a grid).
+class ValuedPoint a where
+    getPoint:: a -> P2 Double
 
--- | insert a new (value,point) in a grid, but check the 9 neighbouring cells for any
--- points touching the new point.  If a touching point is found, a (Left v) is returned,
--- where v was associated with the (first found) touching point.
+-- | A 2D point is considered a ValuedPoint (with no associated value)
+instance ValuedPoint (P2 Double) where
+    getPoint p = p
+
+-- | A pair of a value and a 2D point is considered a ValuedPoint
+instance ValuedPoint (a, P2 Double) where
+    getPoint ((_,p)) = p
+
+-- | get the Int x and Int y grid coordinates for a ValuedPoint (using floor).
+gridCoords :: ValuedPoint a => a -> (Int,Int)
+gridCoords a = (n,m) 
+   where p = getPoint a
+         !n = floor (p ^. _x) 
+         !m = floor (p ^. _y)
+
+-- | insert a new valued point in a grid, but check the 9 neighbouring cells for any
+-- points touching the new point.  If a touching point is found, a (Left vpt) is returned,
+-- where vpt is the (first found) grid entry with a touching point.
 -- Otherwise Right gd' is returned where gd' is an updated grid with the new (valued) point.
-insertGridCheck :: (a, P2 Double) -> Grid a -> Either a (Grid a)
-insertGridCheck ap@(_,p) gd =
-    case checkGridNear p gd of
+insertGridCheck :: ValuedPoint a => a -> Grid a -> Either a (Grid a)
+insertGridCheck a gd =
+    case checkGridNear (getPoint a) gd of
         Just b  -> Left b
-        Nothing -> Right $ insertGrid ap gd
+        Nothing -> Right $ insertGrid a gd
 
- -- | insert a new (valued) point in a grid without any checks.
- -- This is used to initialise a grid with (valued) where the points are known not to be touching.       
-insertGrid :: (a, P2 Double) -> Grid a -> Grid a
+-- | insert a new (valued) point in a grid without any checks.
+-- This is used to initialise a grid with (valued) points where the points are known not to be touching.       
+insertGrid :: ValuedPoint a => a -> Grid a -> Grid a
 insertGrid ap gd = Grid $ IMap.alter column n (gridmap gd)
-   where (n,m) = gridCoords (snd ap)
+   where (n,m) = gridCoords ap
          column Nothing = Just $ IMap.insert n [ap] IMap.empty
          column (Just imp) = Just $ IMap.alter ht m imp
          ht Nothing = Just [ap]
          ht (Just aps) = Just (ap:aps)
 
--- | get the list of (valued) points in a grid cell (with Int coords x y)         
-fromGrid :: Grid a -> Int -> Int -> [(a,P2 Double)]
+-- | get the list of (valued) points in a grid cell (with the given Int coords)         
+fromGrid :: Grid a -> Int -> Int -> [a]
 fromGrid gd n = 
     case IMap.lookup n (gridmap gd) of
       Nothing ->  (\_ -> [])
@@ -77,8 +91,8 @@ fromGrid gd n =
                           Nothing -> []
                           Just aps -> aps)
 
--- | get the list of (valued) points from 9 grid cells (aound the one with Int coords x y)         
-fromGridNear :: Grid a -> Int -> Int -> [(a,P2 Double)]
+-- | get the list of (valued) points from 9 grid cells (aound the one with given Int coords)         
+fromGridNear :: Grid a -> Int -> Int -> [a]
 fromGridNear gd n m =  
     let g0 = fromGrid gd n
         g1 = fromGrid gd (n-1)
@@ -90,12 +104,12 @@ fromGridNear gd n m =
 
 -- | check a point for any touching locations in the grid
 -- (checking with points in the 9 grid cells near the point).
--- This will return Nothing if there are no touching locations
--- and Just v otherwise where v is the value paired with the first touching location found.       
-checkGridNear :: P2 Double -> Grid a -> Maybe a             
-checkGridNear p gd = fmap fst $ find touchingItem (fromGridNear gd n m) where
-                 (n,m) = gridCoords p
-                 touchingItem (_,p') = touching p p'
+-- This will return Nothing if there are no touching locations and Just vpt otherwise
+-- where vpt is the the first valued point found in the grid with a touching location.       
+checkGridNear :: ValuedPoint a => P2 Double -> Grid a -> Maybe a             
+checkGridNear p gd = find touchingItem (fromGridNear gd n m)
+  where (n,m) = gridCoords p
+        touchingItem a = touching p (getPoint a)
 
 {-|touching checks if two points are considered close.
 Close means the square of the distance between them is less than a certain number (currently 0.25) so they cannot be
@@ -109,18 +123,13 @@ touching p p1 = quadrance (p .-. p1) < 0.25 -- quadrance is square of length of 
                  
 -- | insert (valued) points into a new grid without any checks.
 -- This is used to initialise a grid with (valued) points known not to be touching.       
-createGrid :: [(a,P2 Double)] -> Grid a
+createGrid :: ValuedPoint a => [a] -> Grid a
 createGrid aps = foldl' (flip insertGrid) (Grid IMap.empty) aps
 
--- | insert points into a new grid without any checks.
--- This uses () for values.      
-createPointGrid :: [P2 Double] -> Grid ()
-createPointGrid = createGrid . map ((),) 
-
 -- | insert (valued) points into a new grid with checks.
--- This will return Left v if there is a touching location with value v,
+-- This will return Left vpt if vpt is found in the grid with a touching location,
 -- and Just gd otherwise where gd is a grid containiing the (valued) points.
-createGridCheck :: [(a, P2 Double)] -> Either a (Grid a)
+createGridCheck :: ValuedPoint a => [a] -> Either a (Grid a)
 createGridCheck aps = insertAll aps (Grid IMap.empty) where
   insertAll [] gd = Right gd
   insertAll (ap:more) gd =